Add Terraform installation and configuration updates across deployment scripts
- Introduced Terraform installation steps in pipeline YAML files.
- Enhanced helper scripts to manage Terraform installation and versioning.
- Updated Azure login process to support Managed Service Identity.
- Added Ansible version configuration file.

+++ deploy/pipelines/helper.sh
@@ -49,20 +49,26 @@ function configureNonDeployer() {
 
 	sudo apt-get -qq install zip
 
-	echo -e "$green --- Install terraform ---$reset"
+	if ! which terraform; then
+		if [ -n "$tf_version" ]; then
+			echo -e "$green--- Install Terraform version $tf_version ---$reset"
+			tf_url="https://releases.hashicorp.com/terraform/${tf_version}/terraform_${tf_version}_linux_amd64.zip"
+		else
+			echo -e "$green--- Install latest Terraform ---$reset"
+			tf_version=$(curl -s https://checkpoint-api.hashicorp.com/v1/check/terraform | jq -r '.current_version')
+			tf_url="https://releases.hashicorp.com/terraform/${tf_version}/terraform_${tf_version}_linux_amd64.zip"
+		fi
 
-	wget -q "$tf_url"
-	return_code=$?
-	if [ 0 != $return_code ]; then
-		echo "##vso[task.logissue type=error]Unable to download Terraform version $tf_version."
-		exit 2
+		wget -q "$tf_url"
+		return_code=$?
+		if [ 0 != $return_code ]; then
+			echo "##vso[task.logissue type=error]Unable to download Terraform version $tf_version."
+			exit 2
+		fi
+		unzip -qq "terraform_${tf_version}_linux_amd64.zip"
+		sudo mv terraform /bin/
+		rm -f "terraform_${tf_version}_linux_amd64.zip"
 	fi
-	unzip -qq "terraform_${tf_version}_linux_amd64.zip"
-	sudo mv terraform /bin/
-	rm -f "terraform_${tf_version}_linux_amd64.zip"
-
-	az extension add --name storage-blob-preview --allow-preview true --output none >/dev/null
-
 }
 
 function LogonToAzure() {
@@ -76,13 +82,15 @@ function LogonToAzure() {
 		az login --service-principal --client-id "$ARM_CLIENT_ID" --password="$ARM_CLIENT_SECRET" --tenant "$ARM_TENANT_ID" --output none
 	else
 		echo "Deployment credentials:              Managed Service Identity"
-		source "/etc/profile.d/deploy_server.sh"
+		if [[ -f /etc/profile.d/deploy_server.sh ]]; then
+			source "/etc/profile.d/deploy_server.sh"
 
-		# sourcing deploy_server.sh overwrites ARM_SUBSCRIPTION_ID with control plane subscription id
-		# ensure we are exporting the right ARM_SUBSCRIPTION_ID when authenticating against workload zones.
-		if [[ "$ARM_SUBSCRIPTION_ID" != "$subscriptionId" ]]; then
-			ARM_SUBSCRIPTION_ID=$subscriptionId
-			export ARM_SUBSCRIPTION_ID
+			# sourcing deploy_server.sh overwrites ARM_SUBSCRIPTION_ID with control plane subscription id
+			# ensure we are exporting the right ARM_SUBSCRIPTION_ID when authenticating against workload zones.
+			if [[ "$ARM_SUBSCRIPTION_ID" != "$subscriptionId" ]]; then
+				ARM_SUBSCRIPTION_ID=$subscriptionId
+				export ARM_SUBSCRIPTION_ID
+			fi
 		fi
 	fi
